Add files via upload
My goal is to have data validation  in columns 6 and 7 for each google sheet. The code worked for the first sheet, but I don't know how to write the code to have it distinguish each sheet separately. The code only seems to work for Menahga, and the last school, and Pequot Lakes.